% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arkrvu-package.R
\docType{package}
\name{arkrvu-package}
\alias{arkrvu}
\alias{arkrvu-package}
\title{arkrvu: Archive: Physician Fee Schedule}
\description{
Archive of Medicare's Physician Fee Schedule.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/andrewallenbruce/arkrvu}
  \item \url{https://andrewallenbruce.github.io/arkrvu/}
  \item Report bugs at \url{https://github.com/andrewallenbruce/arkrvu/issues}
}

}
\author{
\strong{Maintainer}: Andrew Bruce \email{andrewallenbruce@gmail.com} [copyright holder]

}
\keyword{internal}
