% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{conv_fct}
\alias{conv_fct}
\title{Conversion Factor}
\format{
\subsection{\code{conv_fct}}{

A data frame with 36 rows and 6 columns:
\describe{
\item{date_start}{Effective Date}
\item{date_end}{End Date}
\item{date_iv}{Date Interval}
\item{cf}{Conversion Factor}
\item{chg_abs}{Absolute Change}
\item{chg_rel}{Relative Change}
...
}
}
}
\usage{
conv_fct
}
\description{
Conversion Factor
}
\keyword{internal}
