% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf.R
\name{conversion_factor}
\alias{conversion_factor}
\title{Conversion Factor}
\usage{
conversion_factor(dos = NULL)
}
\arguments{
\item{dos}{\verb{<Date>} Date of Service, in the form YYYY-MM-DD; default is \code{NULL}}
}
\value{
\code{tibble} containing conversion factors; if \code{dos} is provided, returns conversion factor for that date
}
\description{
Conversion Factor
}
\examples{
conversion_factor("2023-03-03")
conversion_factor(make_date(2018:2025))
conversion_factor()

}
