% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\name{download_links}
\alias{download_links}
\title{RVU File Download Links}
\usage{
download_links()
}
\value{
list of selected rvu source files
}
\description{
RVU File Download Links
}
\examples{
download_links()

}
