% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloading.R
\name{download_rvu_link_table}
\alias{download_rvu_link_table}
\title{Download RVU Link Table}
\usage{
download_rvu_link_table()
}
\value{
\verb{<tibble>} containing year, file, and url of RVU source files
}
\description{
Downloads the table of RVU source file links from the CMS website. Used
internally to populate and update the \code{rvu_link_table} dataset.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
download_rvu_link_table()
\dontshow{\}) # examplesIf}
}
\keyword{internal}
