% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloading.R
\name{download_rvu_zip_links}
\alias{download_rvu_zip_links}
\title{Download RVU Zip File Link Table}
\usage{
download_rvu_zip_links(years, urls)
}
\arguments{
\item{years}{\verb{<int>} year(s) of RVU source file, available in \code{rvu_link_table()}}
}
\value{
\verb{<tibble>} containing year, file, and url of RVU source files
}
\description{
Downloads the table of RVU source zip file links from the CMS website. Used
internally to populate and update the \code{rvu_zip_links} dataset.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
download_rvu_zip_links(years = 2024)
\dontshow{\}) # examplesIf}
}
\keyword{internal}
