% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\name{download_table}
\alias{download_table}
\title{Download RVU Link Table}
\usage{
download_table()
}
\value{
\verb{<tibble>} containing year, file, and url of RVU source files
}
\description{
Download RVU Link Table
}
\keyword{internal}
