% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\name{download_zip_links}
\alias{download_zip_links}
\title{Download RVU Zip File Links}
\usage{
download_zip_links(years)
}
\arguments{
\item{years}{\verb{<int>} years of RVU source files}
}
\value{
\verb{<tibble>} containing year, file, and url of RVU source files
}
\description{
Download RVU Zip File Links
}
\examples{
\dontshow{if (rlang::is_interactive()) withAutoprint(\{ # examplesIf}
download_zip_links(years = 2024)
\dontshow{\}) # examplesIf}
}
\keyword{internal}
