% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rvu_files.R
\name{get_conversion_factor}
\alias{get_conversion_factor}
\title{Get Conversion Factors by Effective Date}
\usage{
get_conversion_factor(dos = NULL)
}
\arguments{
\item{dos}{\verb{<Date>} date of service}
}
\value{
list of selected rvu source files
}
\description{
Get Conversion Factors by Effective Date
}
\examples{
get_conversion_factor(dos = as.Date("2023-03-03"))

get_conversion_factor() |>
   print(n = 30)

}
