% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rvu_files.R
\name{get_link_table}
\alias{get_link_table}
\title{Get RVU Link Table}
\usage{
get_link_table()
}
\value{
list of selected rvu source files
}
\description{
Get RVU Link Table
}
\examples{
get_link_table()

}
