% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_pin}
\alias{get_pin}
\title{Load \link[pins:pins-package]{pins}}
\usage{
get_pin(pin, ...)
}
\arguments{
\item{pin}{\verb{<chr>} pin name, see \code{\link[=list_pins]{list_pins()}}}

\item{...}{additional arguments passed to \code{\link[=mount_board]{mount_board()}}}
}
\value{
\verb{<tibble>} containing pinned dataset
}
\description{
Load \link[pins:pins-package]{pins}
}
\keyword{internal}
