% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rvu_files.R
\name{get_pprrvu_}
\alias{get_pprrvu_}
\title{Parallelized \code{\link[=get_pprrvu]{get_pprrvu()}}}
\usage{
get_pprrvu_(df, ...)
}
\arguments{
\item{df}{\verb{<tibble>} containing dos, hcpcs, pos}

\item{...}{Pass arguments to \code{\link[=get_pprrvu]{get_pprrvu()}}.}
}
\value{
\verb{<tibble>} containing pprrvu source file
}
\description{
Parallelized \code{\link[=get_pprrvu]{get_pprrvu()}}
}
