% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rvu_files.R
\name{get_source}
\alias{get_source}
\title{Get RVU Files}
\usage{
get_source(
  source = c("link", "zip", "file", "pprvu", "oppscap", "gpci", "locco", "anes")
)
}
\arguments{
\item{source}{rvu source (link, zip, file, pprvu, oppscap, gpci, locco, anes)}
}
\value{
list of selected rvu source files
}
\description{
Get RVU Files
}
\examples{
get_source("link")

get_source("zip")

get_source("file")

get_source("pprvu")

get_source("gpci")

}
