% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcpcs.R
\name{hcpcs_checks}
\alias{hcpcs_checks}
\alias{is_hcpcs}
\alias{is_hcpcs_level_I}
\alias{is_hcpcs_level_II}
\alias{is_cpt_category_I}
\alias{is_cpt_category_II}
\alias{is_cpt_category_III}
\alias{hcpcs_category}
\alias{cpt_category}
\alias{hcpcs_level}
\alias{cpt_level}
\title{Validate HCPCS Codes}
\source{
\href{https://www.ama-assn.org/practice-management/cpt/category-ii-codes}{AMA Category II Codes}

\href{https://www.ama-assn.org/practice-management/cpt/category-iii-codes}{AMA Category III Codes}
}
\usage{
is_hcpcs(hcpcs)

is_hcpcs_level_I(hcpcs)

is_hcpcs_level_II(hcpcs)

is_cpt_category_I(hcpcs)

is_cpt_category_II(hcpcs)

is_cpt_category_III(hcpcs)

hcpcs_category(hcpcs)

cpt_category(hcpcs)

hcpcs_level(hcpcs)

cpt_level(hcpcs)
}
\arguments{
\item{hcpcs}{\verb{<chr>} vector of possible HCPCS codes}
}
\value{
\verb{<lgl>} \code{TRUE} if valid, otherwise \code{FALSE}
}
\description{
Validate HCPCS Codes
}
\section{HCPCS Codes}{

\itemize{
\item 5 characters long
\item Begin with one of \verb{[A-CEGHJ-MP-V0-9]}
\item Followed by any 3 digits, \verb{[0-9]\{3\}}
\item End with one of \verb{[AFMTU0-9]}
}
}

\section{HCPCS Level I (CPT) Codes}{

\itemize{
\item Begin with any 4 digits, \verb{[0-9]\{4\}}
\item End with one of \verb{[AFMTU0-9]}
}

HCPCS Level I is comprised of CPT (Current Procedural Terminology), a uniform
numeric coding system maintained by the AMA, consisting of terms and codes
that are used to identify medical services and procedures furnished by physicians
and other health care professionals, to bill public or private health insurance
programs. Level I does not include codes needed to separately report medical
items or services that are regularly billed by suppliers other than physicians.
}

\section{HCPCS Level II Codes}{

\itemize{
\item Begin with one of \verb{[A-CEGHJ-MP-V]}
\item Ends with any 4 digits, \verb{[0-9]\{4\}}
}

HCPCS Level II is a standardized coding system that is used primarily to
identify products, supplies, and services not included in the Level I CPT
codes, including ambulance services, durable medical equipment, prosthetics,
orthotics, and supplies (DMEPOS) when used outside a physician's office.

Level II codes are also referred to as alpha-numeric codes because they
consist of a single alphabetical letter followed by 4 numeric digits, while
CPT codes are identified using 5 numeric digits.
}

\section{CPT Category I Codes}{

\itemize{
\item Begins with any 4 digits, \verb{[0-9]\{4\}}, and
\item Ends with one of \verb{[AMU0-9]}
}
}

\section{CPT Category II Codes}{

\itemize{
\item Begins with any 4 digits, \verb{[0-9]\{4\}}, and
\item Ends with an \verb{[F]}
}

Category II codes are supplemental tracking codes that are used to track
services on claims for performance measurement and are not billing codes.

These codes are intended to facilitate data collection about quality of care
by coding certain services and/or test results that support performance
measures and that have been agreed upon as contributing to good patient care.

Some codes in this category may relate to compliance by the health care
professional with state or federal law. The use of these codes is optional.
The codes are not required for correct coding and may not be used as a
substitute for Category I codes.

Services/procedures or test results described in this category make use of
alpha characters as the 5th character in the string (i.e., 4 digits followed
by an alpha character).

These digits are not intended to reflect the placement of the code in the
regular (Category I) part of the CPT code set.

Also, these codes describe components that are typically included in an
evaluation and management service or test results that are part of the
laboratory test/procedure. Consequently, they do not have a relative value
associated with them.
}

\section{CPT Category III Codes}{

\itemize{
\item Begins with any 4 digits, \verb{[0-9]\{4\}}, and
\item Ends with a \verb{[T]}
}

CPT Category III codes are a set of temporary codes that allow data
collection for emerging technologies and are used in the the Food and Drug
Administration (FDA) approval process.

No RVUs are assigned to these codes and payment is based on payer policy. If
a Category III code is available, it must be reported in place of a Category
I unlisted code.

The procedure/service they describe may not meet the following Category I
requirements:
\itemize{
\item Necessary devices/drugs have received FDA clearance/approval.
\item It is performed by many US physicians/QHPs, at a frequency consistent with intended clinical use.
\item It is consistent with current medical practice.
\item It's clinical efficacy is documented in CPT-approved literature.
}
}

\examples{
x <- c("T1503", "G0478", "81301", "69641", "0583F", "0779T", NA)

is_hcpcs(x)

x[which(is_hcpcs(x))]

try(is_hcpcs("1164"))

is_hcpcs_level_I(x)
is_hcpcs_level_II(x)
is_cpt_category_I(x)
is_cpt_category_II(x)
is_cpt_category_III(x)

fastplyr::new_tbl(
   x = c("39503", "99215", "99140", "69990", "70010",
         "0222U", "V5299", "7010F", "0074T"),
   level = hcpcs_level(x)) |>
collapse::mtt(
    level = cheapr::if_else_(level == "CPT", cpt_level(x), level),
    category = cheapr::if_else_(level != "HCPCS II", cpt_category(x), hcpcs_category(x)))
}
