% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{link_tables}
\alias{link_tables}
\alias{rvu_link_table}
\alias{rvu_zip_links}
\title{RVU File Information Tables}
\usage{
rvu_link_table(years)

rvu_zip_links(years)
}
\arguments{
\item{years}{\verb{<int>} RVU source file year}
}
\value{
\verb{<tibble>} containing information about RVU source files
}
\description{
RVU File Information Tables
}
\examples{
rvu_link_table()

rvu_zip_links()
}
