% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_pins}
\alias{list_pins}
\title{List \link[pins:pins-package]{pins}}
\usage{
list_pins(...)
}
\arguments{
\item{...}{arguments to pass to \code{\link[=mount_board]{mount_board()}}}
}
\value{
\verb{<chr>} vector of named pins
}
\description{
List \link[pins:pins-package]{pins}
}
\keyword{internal}
