% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_date}
\alias{make_date}
\title{Make a date object}
\usage{
make_date(y, m = 1L, d = 1L, ...)
}
\arguments{
\item{y}{\verb{<integer>} year}

\item{m}{\verb{<integer>} month, default is \code{1L}}

\item{d}{\verb{<integer>} day, default is \code{1L}}

\item{...}{additional arguments passed to \code{clock::date_build()}}
}
\value{
\verb{<Date>} date object
}
\description{
Make a date object
}
\examples{
make_date(2015:2025)
make_date(2010, 6, 1)
}
