% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{parse_rvu_zip_links}
\alias{parse_rvu_zip_links}
\title{Process RVU Zip File Link Table}
\usage{
parse_rvu_zip_links(x)
}
\arguments{
\item{x}{\verb{<data.frame>} output of \code{\link[=download_rvu_zip_links]{download_rvu_zip_links()}}}
}
\value{
\verb{<tibble>} containing information about RVU source files
}
\description{
Processes the downloaded table of RVU source zip file links from the CMS
website. Used internally to populate and update the \code{rvu_zip_links} dataset.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
parse_rvu_zip_links(download_rvu_zip_links(2024))
\dontshow{\}) # examplesIf}
}
\keyword{internal}
