% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{place_of_srvc}
\alias{place_of_srvc}
\title{Place of Service}
\format{
\subsection{\code{place_of_srvc}}{

A data frame with 52 rows and 4 columns:
\describe{
\item{pos_code}{Place of Service Code}
\item{pos_type}{Place of Service Type}
\item{pos_name}{Place of Service Name}
\item{pos_description}{Place of Service Description}
...
}
}
}
\usage{
place_of_srvc
}
\description{
Place of Service
}
\keyword{internal}
