% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_source.R
\name{raw_source}
\alias{raw_source}
\title{RVU Raw Source Files}
\usage{
raw_source(year, type)
}
\arguments{
\item{year}{\verb{<int>} year of RVU source file}

\item{type}{\verb{<chr>} RVU source file type, one of:
\itemize{
\item \code{pprrvu}: PPRRVU
\item \code{oppscap}: OPPSCAP
\item \code{gpci}: GPCI
\item \code{locco}: LOCCO
\item \code{anes}: ANESTHESIA
}}
}
\value{
\verb{<list>} of tibbles containing RVU source files
}
\description{
RVU Raw Source Files
}
\examples{
raw_source(year = 2024, type = "pprrvu")
raw_source(year = 2023, type = "gpci")
raw_source(year = 2022)

}
