% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbrvs-package.R
\docType{package}
\name{rbrvs-package}
\alias{rbrvs}
\alias{rbrvs-package}
\title{rbrvs: S7 For Medicare Physician Fee Schedule}
\description{
S7 Classes and Methods for the Components of Medicare's Physician Fee Schedule.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/andrewallenbruce/rbrvs}
  \item \url{https://andrewallenbruce.github.io/rbrvs/}
  \item Report bugs at \url{https://github.com/andrewallenbruce/rbrvs/issues}
}

}
\author{
\strong{Maintainer}: Andrew Bruce \email{andrewallenbruce@gmail.com} [copyright holder]

}
\keyword{internal}
