% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recoding}
\alias{recoding}
\alias{recode_mod}
\alias{recode_glob}
\alias{recode_team}
\alias{recode_bilat}
\alias{recode_mult}
\alias{recode_cosurg}
\alias{recode_asst}
\alias{recode_diag}
\alias{recode_pctc}
\alias{recode_status}
\title{Recode Indicators}
\usage{
recode_mod(x, which = "label")

recode_glob(x, which = "label")

recode_team(x)

recode_bilat(x, which = "label")

recode_mult(x, which = "label")

recode_cosurg(x)

recode_asst(x)

recode_diag(x)

recode_pctc(x, which = "label")

recode_status(x, which = "label")
}
\arguments{
\item{x}{\verb{<chr>} vector of indicators}

\item{which}{\verb{<chr>} which information to return; \code{"label"}, \code{"desc"} or \code{"rule"} (see individual functions for available options)}
}
\value{
\verb{<chr>} vector of descriptions
}
\description{
Recode Indicators
}
\section{Multiple Surgery}{


100\% of MPFS amount is allowed for highest valued surgical procedure and 50\%
for additional surgical procedures (with a multiple surgery indicator of "2")
performed same day. Modifier 51 will be appended to identify reduced
services.
}

\section{Assistant Surgery}{


Modifiers:
\itemize{
\item 80: Assistance by Another Physician
\item 81: Minimal Assistance by a Another Physician
\item 82: Assistance by Another Physician when Qualified Resident Surgeon Unavailable
\item AS: Non-Physician Assistant at Surgery
}

Physician Assistant-at-surgery Services:
\itemize{
\item Allowed at 85\% of MPFS and then 16\% of that amount is allowed for Assistant-at-surgery.
}
}

\section{Diagnostic Imaging}{


Identifies the applicable Diagnostic Service family for HCPCS codes with a
Multiple Procedure indicator of \code{4}.
}

\examples{
# PC/TC Modifiers
recode_mod(c("26", "TC", "53"))

# Global Days
recode_glob(c(0, 1, 9, "M", "Y", "Z"))

# Team Surgery (Mod 66)
recode_team(0:2)

# Bilateral Surgery (Mod 50)
recode_bilat(0:3)

# Multiple Procedure (Mod 51)
recode_mult(0:7)

# Co-Surgeon (Mod 62)
recode_cosurg(0:2)

# Assistant Surgery (Mods 80-82, AS)
recode_asst(0:2)

# Diagnostic Imaging Reduction (mult == 4)
recode_mult(4)
recode_diag(1)
recode_pctc(1)

# PC/TC Indicator
recode_pctc(0:8, "rule")

# Status Codes
recode_status(LETTERS[c(1:10, 13:14, 16, 18, 20, 24)])

}
