% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvu-package.R
\docType{package}
\name{rvu-package}
\alias{rvu}
\alias{rvu-package}
\title{rvu: Tidy Revenue Integrity Tools}
\description{
Revenue integrity tools.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/andrewallenbruce/rvu}
  \item \url{https://andrewallenbruce.github.io/rvu/}
  \item Report bugs at \url{https://github.com/andrewallenbruce/rvu/issues}
}

}
\author{
\strong{Maintainer}: Andrew Bruce \email{andrewallenbruce@gmail.com} [copyright holder]

}
\keyword{internal}
