% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\name{rvu_link_table}
\alias{rvu_link_table}
\title{RVU File Download Links}
\usage{
rvu_link_table(years)
}
\arguments{
\item{years}{\verb{<int>} years of RVU source files}
}
\value{
\verb{<tibble>} containing year, file, and url of RVU source files
}
\description{
RVU File Download Links
}
\examples{
rvu_link_table()

}
