% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvu_lookup.R
\name{rvu_lookup}
\alias{rvu_lookup}
\title{RVU Lookup by Date of Service, HCPCS, and Place of Service}
\usage{
rvu_lookup(dos, hcpcs, pos, ...)
}
\arguments{
\item{dos}{\verb{<date>} date of service; YYYY-MM-DD}

\item{hcpcs}{\verb{<chr>} HCPCS code or vector of HCPCS codes}

\item{pos}{\verb{<chr>} Place of Service indicator; \code{Facility} or \code{Non-Facility}}

\item{...}{additional arguments}
}
\value{
\verb{<tibble>} containing RVU values for specified parameters
}
\description{
RVU Lookup by Date of Service, HCPCS, and Place of Service
}
\examples{
rvu_lookup(dos = "2024-03-31", hcpcs = "99215", pos = "Non-Facility")
rvu_lookup(dos = "2023-02-28", hcpcs = "99215", pos = "Facility")
rvu_lookup(dos = "2022-02-28", hcpcs = "99215", pos = "Facility")
}
