% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\name{rvu_zip_links}
\alias{rvu_zip_links}
\title{RVU Zip File Download Links}
\usage{
rvu_zip_links(years)
}
\arguments{
\item{years}{\verb{<int>} years of RVU source files}
}
\value{
\verb{<tibble>} containing year, file, and date_start, date_end, description, size and url of RVU zip files
}
\description{
RVU Zip File Download Links
}
\examples{
rvu_zip_links()

}
