% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\name{download_rvu_link_table}
\alias{download_rvu_link_table}
\title{Download RVU Link Table}
\usage{
download_rvu_link_table()
}
\value{
\verb{<tibble>} containing year, file, and url of RVU source files
}
\description{
Download RVU Link Table
}
\examples{
\dontshow{if (rlang::is_interactive()) withAutoprint(\{ # examplesIf}
download_rvu_link_table()
\dontshow{\}) # examplesIf}
}
\keyword{internal}
