% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rvu_files.R
\name{get_conversion_factor}
\alias{get_conversion_factor}
\title{Get Conversion Factors by Effective Date}
\usage{
get_conversion_factor()
}
\value{
list of selected rvu source files
}
\description{
Get Conversion Factors by Effective Date
}
\examples{
get_conversion_factor()

}
