% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rvu_files.R
\name{get_pprrvu}
\alias{get_pprrvu}
\title{PPRRVU File by Year}
\usage{
get_pprrvu(dos, hcpcs, pos, ...)
}
\arguments{
\item{dos}{\verb{<date>} date of service; YYYY-MM-DD}

\item{hcpcs}{\verb{<chr>} hcpcs code}

\item{pos}{\verb{<chr>} place of service; \code{Facility} or \code{Non-Facility}}

\item{...}{additional arguments}
}
\value{
\verb{<tibble>} containing PPRRVU source file
}
\description{
PPRRVU File by Year
}
\examples{
get_pprrvu("2024-03-31", "99213", "Non-Facility")
get_pprrvu("2024-02-28", "99213", "Facility")
}
