% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rvu_files.R
\name{get_pprrvu}
\alias{get_pprrvu}
\title{Get PPRRVU File by Year}
\usage{
get_pprrvu(year)
}
\arguments{
\item{year}{\verb{<int>} year of rvu source file; default is \code{2023}}
}
\value{
\verb{<tibble>} containing pprrvu source file
}
\description{
Get PPRRVU File by Year
}
\examples{
get_pprrvu(2024)

get_pprrvu(2023)

}
