% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rvu_files.R
\name{get_source}
\alias{get_source}
\title{Get RVU Files}
\usage{
get_source(year, source)
}
\arguments{
\item{source}{rvu source (link, zip, file, pprvu, oppscap, gpci, locco, anes)}
}
\value{
list of selected rvu source files
}
\description{
Get RVU Files
}
\examples{
get_source(2024, "link")

get_source(2024, "zip")

get_source(2024, "file")

get_source(2024, "pprvu")

get_source(2024, "gpci")

}
