% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rvu_files.R
\name{get_source}
\alias{get_source}
\title{Get RVU Source File}
\usage{
get_source(year, source)
}
\arguments{
\item{year}{\verb{<int>} year of rvu source file; default is \code{2020}}

\item{source}{\verb{<chr>} rvu source file (\code{pprvu}, \code{oppscap}, \code{gpci}, \code{locco}, \code{anes})}
}
\value{
\verb{<list>} of tibbles containing rvu source files
}
\description{
Get RVU Source File
}
\examples{
get_source(2024, "pprrvu")
get_source(2024, "gpci")
}
