% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recoding}
\alias{recoding}
\alias{recode_mod}
\alias{recode_glob}
\alias{recode_team}
\alias{recode_bilat}
\alias{recode_mult}
\alias{recode_cosurg}
\alias{recode_asst}
\alias{recode_diag}
\alias{recode_pctc}
\alias{recode_status}
\alias{recode_podp}
\title{Recoding Indicators}
\usage{
recode_mod(x)

recode_glob(x)

recode_team(x)

recode_bilat(x)

recode_mult(x)

recode_cosurg(x)

recode_asst(x)

recode_diag(x)

recode_pctc(x)

recode_status(x, which = c("name", "description"))

recode_podp(x)
}
\arguments{
\item{x}{\verb{<chr>} vector of indicators}

\item{which}{\verb{<chr>} which recoding to perform; one of \code{"name"} or \code{"description"}}
}
\value{
\verb{<chr>} vector of descriptions
}
\description{
Recoding Indicators
}
\section{Multiple Surgery}{


100\% of MPFS amount is allowed for highest valued surgical procedure and 50\%
for additional surgical procedures (with a multiple surgery indicator of "2")
performed same day. Modifier 51 will be appended to identify reduced
services.
}

\section{Assistant Surgery}{


Fee schedule amount equals 16\% of amount otherwise applicable for surgical payment.

Modifiers:
\itemize{
\item 80: Assistance by Another Physician
\item 81: Minimal Assistance by a Another Physician
\item 82: Assistance by Another Physician when Qualified Resident Surgeon Unavailable
\item AS: Non-Physician Assistant at Surgery
}

Physician Assistant-at-surgery Services:
\itemize{
\item Allowed at 85\% of MPFS and then 16\% of that amount is allowed for Assistant-at-surgery.
}
}

\section{Diagnostic Imaging}{


Identifies the applicable Diagnostic Service family for HCPCS codes with a
Multiple Procedure indicator of \code{4}.
}

\examples{
# Modifier
recode_mod(c("26", "TC", "53", NA))

# Global Days
recode_glob(c("000", "010", "090", "MMM", "XXX", "YYY", "ZZZ"))

# Team Surgery (Mod 66)
recode_team(c(0:2, "9"))

# Bilateral Surgery (Mod 50)
recode_bilat(c(0:3, "9"))

# Multiple Procedure (Mod 51)
recode_mult(c(0:7, "9"))

# Co-Surgeon (Mod 62)
recode_cosurg(c(0:2, "9"))

# Assistant Surgery (Mods 80-82, AS)
recode_asst(c(0:2, "9"))

# Diagnostic Imaging Reduction (mult == `4`)
recode_diag(c("88", "99"))

# PC/TC Indicator
recode_pctc(c(0:8, "9"))

# Status Codes
recode_status(LETTERS)
recode_status(LETTERS, "description")
}
\concept{recode}
